Storage item service tests

+++ MultiAppBackend/src/main/java/com/example/MultiAppBackend/homizer/storageItem/StorageItemService.java
@@ -1,28 +1,52 @@
 package com.example.MultiAppBackend.homizer.storageItem;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class StorageItemService {
 
     private final StorageItemRepo storageItemRepo;
+
     public void saveStorageItem(StorageItem storageItem) {
-        storageItemRepo.save(storageItem);
+        if (storageItemRepo.findById(storageItem.getId()).isEmpty()) {
+            storageItemRepo.save(storageItem);
+        } else {
+            throw new DuplicateKeyException("Duplicated ID: " + storageItem.getId());
+        }
+
     }
 
     public List<StorageItem> getAllStorageItems() {
-        return storageItemRepo.findAll();
+        List<StorageItem> storageItems = storageItemRepo.findAll();
+        if (!storageItems.isEmpty()) {
+            return storageItems;
+        } else {
+            throw new NoSuchElementException("List is empty");
+        }
     }
 
     public StorageItem getOneStorageItem(String id) {
-         return storageItemRepo.findById(id).get();
+        Optional<StorageItem> storageItem = storageItemRepo.findById(id);
+        if (storageItem.isPresent()){
+            return storageItem.get();
+        } else {
+            throw new NoSuchElementException("Item with id: " + id + " not found!");
+        }
     }
 
     public void deleteStorageItem(String id) {
-        storageItemRepo.deleteById(id);
+        Optional<StorageItem> storageItem = storageItemRepo.findById(id);
+        if (storageItem.isPresent()){
+            storageItemRepo.deleteById(id);
+        } else {
+            throw new NoSuchElementException("Item with id: " + id + " not found!");
+        }
     }
 }
